changed file structure so that each component and its corresponding css file live in a separate directory. updated relative import paths in affected files
